# Improves performance of annotations constraint