## Issue #, if available:

None

## Description of changes:

The current implementation of AnnotationsV2 internally converts the simplified syntax to a type definition (standard syntax) in order to simplify the validation logic. However, it is much faster to directly compare the list of annotations on a value to a set of required and/or permitted annotations, so I updated the AnnotationsV2 model so that it can keep the simplified syntax in a more efficient way.

## Performance

### Methodology

I created a set of type definitions and some values to test them against (see below). I ran warmups invocations until the time was stable, then I use an "[autorange](https://docs.python.org/3/library/timeit.html#timeit.Timer.autorange)" approach to determine the number of invocations in each sample, and then collected and averaged the ops/s for 30 samples.

Types:
```
{ annotations: closed::[foo] }
{ annotations: required::[foo] }
{ annotations: required::closed::[foo] }
```
Values:
```
0
foo::1
bar::2
foo::bar::3
```


### Results


|   Case   | Before (ops/s) | After (ops/s) | Difference |
|:--------:|---------------:|--------------:|-----------:|
| required |        1558456 |      14699298 |      +843% |
|  closed  |        1527099 |      23569745 |     +1443% |
|   both   |         956579 |      10226788 |      +969% |


## Related PRs in ion-schema, ion-schema-tests, ion-schema-schemas:

None

----

_By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license._